atualização

+++ FixacaoConceitosBasicos1/Model/CalculaAluguelCarro.cs
@@ -9,8 +9,8 @@ namespace FixacaoConceitosBasicos1.Model
     internal class CalculaAluguelCarro
     {
         public static double valorDiaria = 95;
-        public static double valorQuilometro = 0.35m;
-        public static int Calcular(int dias, int quilometragemInicial, int quilometragemFinal)
+        public static double valorQuilometro = 0.35;
+        public static double Calcular(int dias, int quilometragemInicial, int quilometragemFinal)
         {
            return (CalculaAluguelCarro.valorDiaria * dias) + ((quilometragemFinal - quilometragemInicial) * CalculaAluguelCarro.valorQuilometro);
         }
